requirements.txt
asgiref==3.8.1
Django==5.1.6
numpy==1.23.5
pandas==2.3.3
python-telegram-bot==22.5
pytz==2024.2
Requests==2.32.5
scikit_learn==1.8.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
